feat(config): Refactor configuration commands to use new input and logger packages
- cli/input : Includes all cli/terminal inputs
