--- internal/cli/output.go
@@ -1,14 +1,0 @@
-package cli
-
-// func MonitorListOutput() {
-
-// }
-
-// func AddToOutput(template string) {
-
-// }
-
-// case:
-// read TEMPLATE
-// inject data
-// print to stdout
